requirements.txt
crawl4ai
PyYAML
dateparser
beautifulsoup4
python-dotenv
pypdf
